ml_services/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
google-generativeai==0.8.5
python-dotenv==1.1.1
